refactor: move address hooks to their own folder
